ai-server-2-refactoring/requirements.txt
absl-py==1.4.0
anyio==3.7.0
astunparse==1.6.3
beanie==1.19.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
fastapi==0.98.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
keras<2.13,>=2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-model==0.0.5
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
motor==3.2.0
mpmath==1.3.0
networkx==3.1
numpy>=1.22,<=1.23.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pymongo==4.4.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorboard<2.13,>=2.12
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
Werkzeug==2.3.6
wrapt<1.15,>=1.11.0
